Add files via upload
Analysis 

The School Board is concerned about evidence of academic dishonesty in the Reading and Math grades of Thomas High School ninth grade students. In an effort to uphold and maintain state-testing standards they have requested an analysis to address the problem. To curtail the impact of this possible academic dishonesty they have decided to equate the grades in question to ‘Nan’, this is the equivalence of saying the math and reading grades of Thomas High School ninth graders is equal to a null value. 

Due to above mentioned changes The School Board would like to determine how equating the grades of these students to a null value will impact, affect or create changes in performance, specifically comparatively to other districts and schools’ based on other variables that were previously assessed results when the results had no null values.  The following analysis will seek to address the queries of The School Board. 

Results of the Analysis

	District Analysis
The strategy of equating all ninth - grade students of Thomas High School Math and Reading grades ‘Nan’ suggested by The School Board to address the issue of academic dishonesty had a minor impact on the district analysis. This minute change was less than 1%. This may be due to the student count of ninth grade students at Thomas High School relative to the total student count of 39,170. Hence making adjustments to a small percentage of the sample would not have caused large adjustments in the averages of the results. 

Analysis
•	School Summary
There were no changes in the ranks of the top schools. Notably, Thomas High School even with these changes was not affected.
Even though the ranking of Thomas High School was not affected by these changes, the average Math, Reading and Overall scores were affected. However, this was in a very minute way which was not vast enough to change its ranking as a school in the district.
Neither the Bottom nor Top Schools saw any impact from these changes as the percentage relative to the sample was very small. 

•	Math and Reading Scores 
As expected, this change only affected the Math and Reading grades of the ninth-grade students at Thomas High School. 

•	Scores by School Spending Scores
The Thomas High School spending group which was the $630 -644 per student had a small change in score. 

•	Scores by School Size 
Thomas High School is in the medium size school scores (1000-2000). Hence, this change saw an impact. However, similarly to the other effects of the change in scores it was also relatively small and created very little change.

•	Scores by School Type
The Charter type school saw a low change score because it is the category associated with Thomas High School. Conversely, District type school saw no impact from the changes as Thomas High School was not a District School type and the changes would not affect it. 

  
Summary

1.	District Analysis – School or student count saw no impact from the changes in the scores to Nan.
2.	School Ranking – There were no changes in any of the rankings for both top and bottom schools even though the scores of Thomas High School changed.
3.	School Sizes - The Medium grouping (1000-2000) did experience changes but by a very small percent a Thomas High School was in this category.
4.	School Type – The school type directly associated with Thomas High School; Charter type did see changes to the grouping.  type.



Analysis 

The School Board is concerned about evidence of academic dishonesty in the Reading and Math grades of Thomas High School ninth grade students. In an effort to uphold and maintain state-testing standards they have requested an analysis to address the problem. To curtail the impact of this possible academic dishonesty they have decided to equate the grades in question to ‘Nan’, this is the equivalence of saying the math and reading grades of Thomas High School ninth graders is equal to a null value. 

Due to above mentioned changes The School Board would like to determine how equating the grades of these students to a null value will impact, affect or create changes in performance, specifically comparatively to other districts and schools’ based on other variables that were previously assessed results when the results had no null values.  The following analysis will seek to address the queries of The School Board. 

Results of the Analysis

	District Analysis
The strategy of equating all ninth - grade students of Thomas High School Math and Reading grades ‘Nan’ suggested by The School Board to address the issue of academic dishonesty had a minor impact on the district analysis. This minute change was less than 1%. This may be due to the student count of ninth grade students at Thomas High School relative to the total student count of 39,170. Hence making adjustments to a small percentage of the sample would not have caused large adjustments in the averages of the results. 

Analysis
•	School Summary
There were no changes in the ranks of the top schools. Notably, Thomas High School even with these changes was not affected.
Even though the ranking of Thomas High School was not affected by these changes, the average Math, Reading and Overall scores were affected. However, this was in a very minute way which was not vast enough to change its ranking as a school in the district.
Neither the Bottom nor Top Schools saw any impact from these changes as the percentage relative to the sample was very small. 

•	Math and Reading Scores 
As expected, this change only affected the Math and Reading grades of the ninth-grade students at Thomas High School. 

•	Scores by School Spending Scores
The Thomas High School spending group which was the $630 -644 per student had a small change in score. 

•	Scores by School Size 
Thomas High School is in the medium size school scores (1000-2000). Hence, this change saw an impact. However, similarly to the other effects of the change in scores it was also relatively small and created very little change.

•	Scores by School Type
The Charter type school saw a low change score because it is the category associated with Thomas High School. Conversely, District type school saw no impact from the changes as Thomas High School was not a District School type and the changes would not affect it. 

  
Summary

1.	District Analysis – School or student count saw no impact from the changes in the scores to Nan.
2.	School Ranking – There were no changes in any of the rankings for both top and bottom schools even though the scores of Thomas High School changed.
3.	School Sizes - The Medium grouping (1000-2000) did experience changes but by a very small percent a Thomas High School was in this category.
4.	School Type – The school type directly associated with Thomas High School; Charter type did see changes to the grouping.  type.